Refactor and split Koog properties tests into targeted classes
